App/backend/named-entity-recognition/requirements.txt
colorlog==6.8.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
fastapi==0.110.0
pydantic==2.6.4
pydantic-core==2.16.3
httpx==0.27.0
requests==2.31.0
spacy==3.7.4
spacy-entity-linker==1.0.3
uvicorn==0.29.0
python-dotenv==1.0.1
SPARQLWrapper==2.0.0